% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{assignToGRanges}
\alias{assignToGRanges}
\title{Assign variant IDs to a GRanges object based on overlap}
\usage{
assignToGRanges(variant_ids, targetGRanges, prune = TRUE)
}
\arguments{
\item{variant_ids}{Character vector of variant IDs}

\item{targetGRanges}{GenomicRanges::GRanges object}

\item{prune}{Boolean, if TRUE will only rturn rows that overlap to >=1 variants}
}
\value{
GenomicRanges::GRanges with an additional mcol $variant_ids.
$variant_ids is a IRanges::CharacterList
}
\description{
Assign variant IDs to a GRanges object based on overlap
}
\details{
This function considers strandedness when finding overlap. "A/G"
variants will only be assigned to "+" stranded ranges, and vice versa.
However, you must make sure that only "A/G" and "T/C" variants are
included in the input.
}
\examples{
Refer to `vignette("Adenine Base Editor analysis")`
}
