% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.preprocessing.R
\name{assignVariantIds}
\alias{assignVariantIds}
\title{Assign overlapping VCF variant IDs to a GRanges}
\usage{
assignVariantIds(range, vcf, sep = ";")
}
\arguments{
\item{range}{GenomicRanges::GRanges object}

\item{vcf}{VariantAnnotation::VCF object}

\item{sep}{Separator to use if >1 variants overlap to a range}
}
\value{
Input \code{range} except adding a $variant_ids column
}
\description{
Assign overlapping VCF variant IDs to a GRanges
}
\examples{
vignette("rABE-analysis")
}
