% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{BranchPointScan}
\alias{BranchPointScan}
\title{Distance from branchpoint to 3' splicing site}
\usage{
BranchPointScan(
  GRanges.intron,
  branchpoint.motif,
  BSgenome = BSgenome.Dmelanogaster.UCSC.dm6::BSgenome.Dmelanogaster.UCSC.dm6,
  logodds.threshold = 0.5
)
}
\arguments{
\item{GRanges.intron}{GenomicRanges::GRanges of the introns.}

\item{branchpoint.motif}{An universalmotif::universalmotif-class object
representing the branch point. Can be loaded using universalmotif::read_*
methods. For example, \code{universalmotif::read_meme()}.}

\item{BSgenome}{BSgenome object from Bioconductor.}

\item{logodds.threshold}{logodds threshold used by
\code{universalmotif::scan_sequences()}.}
}
\value{
A numeric vector in order of ranges of GRanges.intron. Ranges that
do not have identified branchpoint will take NA values.
}
\description{
Distance from branchpoint to 3' splicing site
}
\examples{
vignette("intron-properties")
}
