% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{diffVariants}
\alias{diffVariants}
\title{Set difference of two variant tables}
\usage{
diffVariants(varTableX, varTableY)
}
\arguments{
\item{varTableX}{variant table X}

\item{varTableY}{variant table Y}
}
\value{
variant table of X \ Y (based on $variant_id)
}
\description{
Set difference of two variant tables
}
\examples{
vignette("rABE-analysis")
}
