% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{.MaxEntScanRun}
\alias{.MaxEntScanRun}
\title{System cell of MaxEntScan perl script}
\usage{
.MaxEntScanRun(ps.MES, sequences, ME.dir)
}
\arguments{
\item{ps.MES}{What MaxEntScan script to run.}

\item{sequences}{character vector of sequences.}

\item{ME.dir}{Path to the unzipped MaxEntScan directory.}
}
\value{
numeric vector of the MaxEntScan return values.
}
\description{
System cell of MaxEntScan perl script
}
