% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{.strandedShift}
\alias{.strandedShift}
\title{Stranded version of GenomicRanges::shift}
\usage{
.strandedShift(GRanges.input, shift)
}
\arguments{
\item{GRanges.input}{GenomicRanges::GRanges}

\item{shift}{how many bp to shift each range. For '+' stranded range,
a positive shift will be towards 3' of the genomic coordinate. For '-',
a positive shift will be towards 5' of the genome.}
}
\value{
Shifted GenomicRanges::GRanges.
}
\description{
Stranded version of GenomicRanges::shift
}
