% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.preprocessing.R
\name{filterAdenineEditVariants}
\alias{filterAdenineEditVariants}
\title{Filter variants to include only Adenine base editing}
\usage{
filterAdenineEditVariants(assignedRange, stranded = TRUE, pruning = TRUE)
}
\arguments{
\item{assignedRange}{GenomicRanges::GRanges with $variant_ids column.
You can assign variants with the function \code{assignVariantIds()}.}

\item{stranded}{Boolean whether to consider strandedness. If TRUE,
A/G editing for '+' strand, T/C for '-' strand, raising error if any strand
takes '*' value.}

\item{pruning}{Whether to prune values in $variant_ids column. If TRUE,
$variant_ids will only retain adenine base editing events. If FALSE,
$variant_ids is not changed.}
}
\value{
GenomicRanges::GRanges containing ranges with at least one adenine
base editing event.
}
\description{
Filter variants to include only Adenine base editing
}
\examples{
Refer to `vignette("Adenine Base Editor analysis")`
}
