% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexTableReader.R
\name{flexTableReader}
\alias{flexTableReader}
\title{Read 'table' file with flexible file extension}
\usage{
flexTableReader(col_names, col_types, skip, ...)
}
\arguments{
\item{col_names}{Column names, character vector}

\item{col_types}{Column types, same as \code{readr::read_csv()}. See details.}

\item{skip}{Number of lines to skip before reading data}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Reader function that accepts a file path
}
\description{
Read 'table' file with flexible file extension
}
\details{
Depending on the actual file format provided by the user, column types may
not be respected. For example, if the file is excel spreadsheet, then the
column types provided is mapped to that supported by the \code{readxl} package.
}
\examples{
# TODO
}
