% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{getConfidentVariantIds}
\alias{getConfidentVariantIds}
\title{Get 'confident' variants for downstream analysis}
\usage{
getConfidentVariantIds(variantTableList, print.stats = TRUE, min.occurence = 2)
}
\arguments{
\item{variantTableList}{A list of variant tables. Each variant table must have
$variant_id column.}

\item{print.stats}{Boolean, whether to print statistics.}

\item{min.occurence}{Integer, minimum number of occurences for a variant to
be considered as 'confident'.}
}
\value{
A character vector of 'confident' variant IDs.
}
\description{
Get 'confident' variants for downstream analysis
}
\examples{
Refer to `vignette("Adenine Base Editor analysis")`
}
