% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{getNameVarTable}
\alias{getNameVarTable}
\title{Convert GRanges with assigned variants into a tibble}
\usage{
getNameVarTable(assignedGRanges)
}
\arguments{
\item{assignedGRanges}{an "assigned GRanges"
(i.e., GenomicRanges::GRanges with names and \code{variant_ids}).}
}
\value{
A tibble::tibble with columns: \itemize{
\item name, character of \code{name(assignedGRanges)}
\item variant_ids, list(character) of assignedGRanges$variant_ids
}
}
\description{
Convert GRanges with assigned variants into a tibble
}
\examples{
vignette("rABE-analysis")
}
