% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexTableReader.R
\name{getNormalizedParams}
\alias{getNormalizedParams}
\title{Get specific parameters for a table reader}
\usage{
getNormalizedParams(ext, ...)
}
\arguments{
\item{ext}{File extension.}

\item{...}{Parameters, see details.}
}
\value{
Named list of normalized parameters.
}
\description{
Get specific parameters for a table reader
}
\details{
Currently the following parameters are supported: col_names, col_types,
skip. For excel files, col_types is not supported yet.
}
\examples{
# Not exported
}
