% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.preprocessing.R
\name{getPrimaryVariantTable}
\alias{getPrimaryVariantTable}
\title{Get data frame of primary variants from a VCF file}
\usage{
getPrimaryVariantTable(vcf, variant_ids)
}
\arguments{
\item{vcf}{VariantAnnotation::VCF object}

\item{variant_ids}{Identifiers to subset the VCF object}
}
\value{
A data.frame of counts for primary variants of each ID queried.
}
\description{
Get data frame of primary variants from a VCF file
}
