% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexTableReader.R
\name{getTableReader}
\alias{getTableReader}
\title{Get filetype-specific table reader}
\usage{
getTableReader(ext)
}
\arguments{
\item{ext}{File extension.}
}
\value{
A suitable reader function.
}
\description{
Get filetype-specific table reader
}
\examples{
# Not exported.
}
