% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.preprocessing.R
\name{getVariantTable}
\alias{getVariantTable}
\title{Return detailed variant information (mutation rate, counts for REF/ALT).}
\usage{
getVariantTable(vcf, assignedRange)
}
\arguments{
\item{vcf}{VariantAnnotation::VCF object}

\item{assignedRange}{GenomicRanges::GRanges with $variant_ids column.
You can assign variants with the function \code{assignVariantIds()}.}
}
\value{
A tibble::tibble
}
\description{
Return detailed variant information (mutation rate, counts for REF/ALT).
}
\examples{
vignette("rABE-analysis")
}
