% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{grouper}
\alias{grouper}
\title{Grouping of data frame without taking up the ellipsis}
\usage{
grouper(.data, var.group, ...)
}
\arguments{
\item{.data}{Data frame to perform \code{dplyr::group_by()}.}

\item{var.group}{\strong{Quoted} group variable(s). Multiple variables must be
inside \code{gvars()}. To quote group variables use \code{rlang::enexpr()} in the
caller function.}

\item{...}{Must be empty.}
}
\value{
Grouped data frame.
}
\description{
Motivation: the original \code{dplyr::group_by()} uses ellipsis to allow
flexibility in specifying grouping variables. However, sometimes a function
might need the ellipsis for other purposes. In those cases, it is desirable
to "save" the ellipsis by allowing the user to provide all grouping variables
in a single parameter.
}
\details{
This function is not intended for end users.
}
\examples{
# Correct use
f <- function(df, var.group){
    var.group <- rlang::enexpr(var.group)
    return(grouper(df, var.group))
}
# Single group variable
f(mtcars, cyl)
# Multiple group variable
f(mtcars, gvars(cyl,vs))

# Wrong use
f <- function(df, var.group){ return(grouper(df, var.group)) }
# f(mtcars, cyl) will error
}
