% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{MaxEntScan}
\alias{MaxEntScan}
\title{Splicing site strength scoring by MaxEntScan}
\usage{
MaxEntScan(path.zip.MES = "burgelab.maxent.zip", BSgenome, GRange.intron)
}
\arguments{
\item{path.zip.MES}{Path to the MaxEntScan perl program (zipped). While not
provided by the package, you may obtain a copy from the \href{http://hollywood.mit.edu/burgelab/maxent/Xmaxentscan_scoreseq.html}{original MaxEntScan author}.
Alternatively, you may get an archived copy from \href{https://github.com/yeyuan98/archived_external_resources}{Github}.}

\item{BSgenome}{BSgenome object from Bioconductor.}

\item{GRange.intron}{GenomicRanges::GRanges of the introns. Genome must match
that of the BSgenome object. This intron ranges must be stranded (i.e.,
only contains '+' and '-' strand values.)}
}
\value{
data.frame with the following columns. Rows are guaranteed to match
order of the input introns. \itemize{
\item MaxEnt.5ss, score for the 5' splicing site
\item MaxEnt.3ss, score for the 3' splicing site
}
}
\description{
Splicing site strength scoring by MaxEntScan
}
\examples{
vignette("intron-properties")
}
