% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{mergeVariants}
\alias{mergeVariants}
\title{Merge a list of variant tables into one table.}
\usage{
mergeVariants(variantTableList, variantIDsToKeep)
}
\arguments{
\item{variantTableList}{list of variant tables.}

\item{variantIDsToKeep}{which variant IDs to keep in the merged frame}
}
\value{
A single tibble::tibble of merged variant tables with following columns:
\itemize{
\item variant_id, character vector
\item variant_data list of tibble::tibble, each with columns $countREF,
$countALT, $mutatin.perc.
}
}
\description{
Merge a list of variant tables into one table.
}
\examples{
vignette("rABE-analysis")
}
