% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaRcrd.R
\name{metaRcrd}
\alias{metaRcrd}
\title{\code{metaRcrd} record-style vector}
\usage{
metaRcrd(fields, meta.fields, ...)
}
\arguments{
\item{fields}{A named list or data.frame where each row is a record. Names
of this list are the field names for the record vector.}

\item{meta.fields}{A character vector giving fields that should be considered
as "metadata" fields.}

\item{...}{For future extensions. Must be empty.}
}
\value{
An S3 vector of class \code{y3628_metaRcrd}.
}
\description{
This creates a vector where each item is a record of data and metadata fields.
Data fields are used for record equality/comparison operations.
Metadata fields are conceptually "attributes" attached to the data fields.
}
\section{Note}{
Set operations must use \code{vctrs} methods (e.g., \code{\link[vctrs:vec-set]{vctrs::vec_set_union()}}).
Base set operations are not generic and hence invalid.
}

\examples{
require(vctrs)
require(tibble)

## Representing metadata of experimental samples

today <- Sys.Date()
dates <- c(today, today, today, today+1) # when
conds <- factor(c("L", "M", "H", "L"), levels = c("L", "M", "H")) # condition
who <- c("me", "me", "me", "Ahri") # personnel
eg1 <- metaRcrd(list(date=dates, condition=conds, personnel=who), "personnel")

dates <- c(today) # when
conds <- factor("L", levels = c("L", "M", "H")) # condition
who <- "Ahri" # personnel
eg2 <- metaRcrd(list(date=dates, condition=conds, personnel=who), "personnel")

### concatenate
eg_full <- vec_c(eg1, eg2) # c() works just fine too
eg_full

### equality/set operation
eg1 == eg2
vec_set_difference(eg1, eg2) # use vec_set_* methods

### comparison/sort
eg1 <= eg2
sort(eg_full)

### use in a tibble

proj <- c("fancy project 1", "ok project 2") # what project
rec <- list(eg1, eg2) # experiment record for each project
project_record <- tibble(project = proj, record = rec)
project_record

}
\seealso{
\code{\link[vctrs:new_rcrd]{vctrs::new_rcrd()}}
}
