% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mR_filter}
\alias{mR_filter}
\title{Filtering metaRcrd objects}
\usage{
mR_filter(mRcrd, what)
}
\arguments{
\item{mRcrd}{An S3 vector of class \code{y3628_metaRcrd}.}

\item{what}{Filtering criteria. Quoted and evaluated to subset the record.}
}
\value{
Subsetted \code{y3628_metaRcrd}.
}
\description{
Filtering metaRcrd objects
}
\examples{
require(datasets)
my_mtcars <- datasets::mtcars[c("cyl", "hp", "mpg")]
my_mtcars <- metaRcrd(my_mtcars, meta.fields = "mpg")
mR_filter(my_mtcars, mpg > 20)
}
