% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{phastCons}
\alias{phastCons}
\title{Conservation scoring by phastCons}
\usage{
phastCons(GRange.intron, bw.phastCons.path, bed.phastCons.path)
}
\arguments{
\item{GRange.intron}{GenomicRanges::GRanges of the introns.}

\item{bw.phastCons.path}{Path to a bigwig file of phastCons scores. This
may be retrieved from the UCSC (e.g., \href{https://hgdownload.soe.ucsc.edu/goldenPath/dm6/phastCons124way/}{dm6}). Look for
the \code{dm6.phastCons124way.bw} file.}

\item{bed.phastCons.path}{Path to a bed file of conserved regions annotated
by phastCons. This may be retrieved from the UCSC (e.g., \href{https://hgdownload.soe.ucsc.edu/goldenPath/dm6/database/}{dm6}). Look for the
\code{phastConsElements124way.txt.gz} file.}
}
\value{
data.frame. Number of rows is the same as number of ranges of
\code{GRange.intron}. Results have the following columns \describe{
\item{mean}{Mean phastCons values over each range}
\item{perc.in.element}{Percentage of bases in
conserved phastCons elements for each range}
}
}
\description{
Conservation scoring by phastCons
}
\details{
This function can be used for compute scores for any GRanges of interest.
}
\examples{
vignette("intron-properties")
}
