% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{prettyFormatSummary}
\alias{prettyFormatSummary}
\title{Format summary of a numeric vector}
\usage{
prettyFormatSummary(nums)
}
\arguments{
\item{nums}{A numeric vector}
}
\value{
A character(1) showing standard summary statistics (i.e.,
0/25/50/mean/75/100 quantiles).
}
\description{
Format summary of a numeric vector
}
