% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.preprocessing.R
\name{pruneAdenineEditVariants}
\alias{pruneAdenineEditVariants}
\title{Prune a GenomicRanges containing variants to retain only A->G edits}
\usage{
pruneAdenineEditVariants(rangeWithVariants, stranded = TRUE)
}
\arguments{
\item{rangeWithVariants}{GenomicRanges::GRanges with $variant_ids.
Each range can contain multiple variant_ids separated by \verb{;}.
Example $variant_ids = "chr2L:1_A/G;chr2L:100_C/A"}

\item{stranded}{Boolean, whether to consider strand of the ranges}
}
\value{
GenomicRanges::GRanges with only A->G edits
}
\description{
Prune a GenomicRanges containing variants to retain only A->G edits
}
