% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{str_split_summary}
\alias{str_split_summary}
\title{Applying summary function to string splits}
\usage{
str_split_summary(string, pattern, summary)
}
\arguments{
\item{string}{Input vector, passed to \code{\link[stringr:str_split]{stringr::str_split()}}.}

\item{pattern}{Pattern to look for, passed to \code{\link[stringr:str_split]{stringr::str_split()}}.}

\item{summary}{Summary function that will be applied to each split.}
}
\value{
Vector same length as input vector
}
\description{
Applying summary function to string splits
}
\examples{
t <- c("0.1,0.3,0.7", "NA", "0.2", "")
str_split_summary(t, ",", \(x) mean(as.numeric(x)))
}
