% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEtools.core.R
\name{variantID2GRanges}
\alias{variantID2GRanges}
\title{Convert A/G variant ID vector into GenomicRanges::GRanges}
\usage{
variantID2GRanges(variant_ids)
}
\arguments{
\item{variant_ids}{Character vector of VCF variant identifier. Example:
"chr2L:1_A/G"}
}
\value{
GenomicRanges::GRanges object of the variants. All widths are one
as this is record of single base mutations. The following mcols are defined:
\itemize{
\item \code{type}, "A/G" (strand = "+") or "T/C" (strand = "-")
to reflect variant type.
\item \code{variant_id}, the original input as record
}
}
\description{
Convert A/G variant ID vector into GenomicRanges::GRanges
}
\details{
Variant type is extracted but NOT checked at all. You must make sure
that only "A/G" and "T/C" are present as input.
}
\examples{
Not applicable. Internal use only.
}
